Viewport now accepts two frames
The Viewport class is responsible for handling the behaviour of the
application related to the poition and existance of two frames.
It also uses a strategy pattern (ViewportDelegate) to inform its
delegate whether an overlap occours between the two frames every time
one of the two gets updated in its position or a delegate reference
changes.

NOTE: For simplicity for now it can only have two frames. Once I get to
a point where the basic functionalities work, I'll move on to the next
step which would be to have an unlimited number of frames.

The main complexity I can think of is performance. The Viewport would
have to check the overlapping between all the frames and it might take
time to come up with an algorithm which is performant.

However, as the design evolves, there might be the case in which there's
no situation like this. In which case I might think of prioritise the
multiple frames capability up.