Converted `position` to `center` property
`position` was a bit misleading as it could be anything, therefore I
wanted to rename to something more meaningful.

At the beginning I thought of `origin` as to being a better name.
However, to make it more useful for the next iterations, I preferred to
update the overlapping logic calculation with something more flexible
which is a `center` property.