Refactored logic for view port elements handling
* Removed Viewport class and its tests. The responsibilities of creating
and managing overlaps now belongs to `ViewportDataSource` which handles
the viewport view content and updates it when rectangles change position
* Improved overlaps calculation algorithm and now the viewport can
handle more than only two rectangles. The complexity is still quadratic
for creation and linear for update. There's room for improvement,
however for now it is good enough to proceed with the implementation of
other capabilities